feat(act): add grantee module for managing grantees
