+++ test/integration/modules/bzz.spec.ts
@@ -1,4 +1,3 @@
-import { assert, expect } from 'chai'
 import { expect as jestExpect } from 'expect'
 import { Readable } from 'stream'
 import * as bzz from '../../../src/modules/bzz'
@@ -22,20 +21,6 @@ describe('modules/bzz', () => {
     const data = 'hello act'
     let publicKey: string
     let batchID: BatchId
-    const grantees = [
-      '02ceff1422a7026ba54ad89967d81f2805a55eb3d05f64eb5c49ea6024212b12e8',
-      '02ceff1422a7026ba54ad89967d81f2805a55eb3d05f64eb5c49ea6024212b12e9',
-      '02ceff1422a7026ba54ad89967d81f2805a55eb3d05f64eb5c49ea6024212b12ee',
-    ]
-    const patchGrantees = {
-      add: ['02ceff1422a7026ba54ad89967d81f2805a55eb3d05f64eb5c49ea6024212b12e8'],
-      revoke: [
-        '02ceff1422a7026ba54ad89967d81f2805a55eb3d05f64eb5c49ea6024212b12e9',
-        '02ceff1422a7026ba54ad89967d81f2805a55eb3d05f64eb5c49ea6024212b12ee',
-      ],
-    }
-
-    const patchGranteesString = JSON.stringify(patchGrantees)
 
     before(async () => {
       const responsePUBK = await http<{ publicKey: string }>(BEE_KY_OPTIONS, {
@@ -83,45 +68,6 @@ describe('modules/bzz', () => {
       const result = await bzz.uploadFile(BEE_KY_OPTIONS, data, batchID, filename, { act: true })
       const requestOptionsOK = actBeeKyOptions(publicKey, result.history_address, '1')
       const dFile = await bzz.downloadFile(requestOptionsOK, result.reference, filename)
-      expect(Buffer.from(dFile.data).toString()).to.eql(data)
-    })
-
-    it('should create grantee list', async function () {
-      const response = await bzz.createGrantees(BEE_KY_OPTIONS, batchID, grantees)
-      expect(response.ref).to.have.lengthOf(128)
-      expect(response.historyref).to.have.lengthOf(64)
-    })
-
-    it('should download grantee list', async function () {
-      const response = await bzz.createGrantees(BEE_KY_OPTIONS, batchID, grantees)
-      const list = await bzz.getGrantees(response.ref, BEE_KY_OPTIONS)
-      expect(list.data).to.have.lengthOf(grantees.length)
-      list.data.forEach((element: string, _index: number) => {
-        assert.isTrue(grantees.includes(element))
-      })
-    })
-
-    it('should patch grantee list', async function () {
-      const filename = 'act-4.txt'
-      const uploadResult = await bzz.uploadFile(BEE_KY_OPTIONS, data, batchID, filename, { act: true })
-
-      const createResponse = await bzz.createGrantees(BEE_KY_OPTIONS, batchID, grantees)
-      await new Promise(resolve => setTimeout(resolve, 1000))
-      const patchResponse = await bzz.patchGrantees(
-        createResponse.ref,
-        uploadResult.history_address,
-        batchID,
-        patchGranteesString,
-        BEE_KY_OPTIONS,
-      )
-      const list = await bzz.getGrantees(patchResponse.ref, BEE_KY_OPTIONS)
-
-      expect(list.data).to.have.lengthOf(1)
-      expect(list.data[0]).to.eql(patchGrantees.add[0])
-
-      const requestOptionsOK = actBeeKyOptions(publicKey, patchResponse.historyref, '1')
-      const dFile = await bzz.downloadFile(requestOptionsOK, uploadResult.reference, filename)
-
       expect(Buffer.from(dFile.data).toString()).to.eql(data)
     })
   })
